requirements.txt
tweepy>=4.14.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0

# For Selenium approach (twitter_scrap1.py)
selenium>=4.15.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.0

# For News Scraper (news_scraper.py, google_news_business_scraper.py)
playwright>=1.40.0
requests>=2.31.0
nest-asyncio>=1.5.0
lxml>=4.9.0
python-dateutil>=2.8.0

# For ISIN Matching (news_with_isin_scraper.py, company_isin_matcher.py)
openai>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# For LangChain-powered ISIN Matching (news_isin_langchain.py, isin.py, isin_matcher.py)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-experimental>=0.0.50
